% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonalBlock.R
\name{diagonalBLock}
\alias{diagonalBLock}
\title{Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups}
\usage{
diagonalBLock(B, S, l, m1 = 0, m2, p)
}
\arguments{
\item{B}{a matrix of size p by p}

\item{S}{a matrix of size p by p}

\item{l}{penalization parameter}

\item{m1}{the node at which the partition occurs}

\item{m2}{not relevant}

\item{p}{maximum number of variables}
}
\value{
graph adjacency B
}
\description{
Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups
}
\examples{
diagonalBLock(B,S,l, m1 = 0, m2, p)

}
