% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generates random multivariate normal data from a random B or provided adjacency}
\usage{
generate_data(n, p, B = NULL)
}
\arguments{
\item{n}{number of samples to generate}

\item{p}{number of dimesnions of data}

\item{adj}{adjacency matrix}
}
\value{
graph adjacency B
}
\description{
Generates random multivariate normal data from a random B or provided adjacency
}
\examples{
generate_data(100,20)

}
