% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDAGs.R
\name{partial2}
\alias{partial2}
\title{Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups}
\usage{
partial2(
  X,
  l,
  a = NULL,
  m1 = NULL,
  m2 = NULL,
  m3 = NULL,
  m4 = NULL,
  m5 = NULL,
  m6 = NULL,
  m7 = NULL,
  m8 = NULL,
  m9 = NULL,
  eps = 10^(-4),
  maxitr = 100,
  init = NULL
)
}
\arguments{
\item{X}{a matrix of size n by p containing n observations an p variables}

\item{l}{penalization parameter}

\item{m1}{the node at which the partition occurs}

\item{m2}{not relevant}

\item{m3}{not relevant}

\item{m4}{not relevant}

\item{m5}{not relevant}

\item{m6}{not relevant}

\item{m7}{not relevant}

\item{m8}{not relevant}

\item{m9}{not relevant}

\item{eps}{tolerance parameter to decide whether algorithm has converved or not}

\item{maxitr}{maximum number of iterations to run before returning output}

\item{init}{initial estimate of graph adjacency B}
}
\value{
graph adjacency B
}
\description{
Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups
}
\examples{
partial2(X = X, l = 2, m1 = 25)

}
