% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDAGs.R
\name{partial5_inc}
\alias{partial5_inc}
\title{Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups}
\usage{
partial5_inc(
  X,
  l,
  a = NULL,
  m1 = NULL,
  m2 = NULL,
  m3 = NULL,
  m4 = NULL,
  m5 = NULL,
  m6 = NULL,
  m7 = NULL,
  m8 = NULL,
  m9 = NULL,
  l1 = NULL,
  l2 = NULL,
  l3 = NULL,
  l4 = NULL,
  l5 = NULL,
  l6 = NULL,
  l7 = NULL,
  l8 = NULL,
  l9 = NULL,
  eps = 10^(-4),
  maxitr = 100,
  init = NULL
)
}
\arguments{
\item{X}{a matrix of size n by p containing n observations an p variables}

\item{l}{penalization parameter}

\item{m1}{the node at which the first partition occurs}

\item{m2}{the node at which the second partition occurs}

\item{m3}{the node at which the third partition occurs}

\item{m4}{the node at which the fourth partition occurs}

\item{m5}{not relevant}

\item{m6}{not relevant}

\item{m7}{not relevant}

\item{m8}{not relevant}

\item{m9}{not relevant}

\item{eps}{tolerance parameter to decide whether algorithm has converved or not}

\item{maxitr}{maximum number of iterations to run before returning output}

\item{init}{initial estimate of graph adjacency B}
}
\value{
graph adjacency B
}
\description{
Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups
}
\examples{
partial5_inc(X = X, l = 2, m1 = 12, m2 = 24, m3 = 36, m4 = 48)

}
