% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDAGs.R
\name{partial9}
\alias{partial9}
\title{Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups}
\usage{
partial9(
  X,
  l,
  a = NULL,
  m1 = NULL,
  m2 = NULL,
  m3 = NULL,
  m4 = NULL,
  m5 = NULL,
  m6 = NULL,
  m7 = NULL,
  m8 = NULL,
  m9 = NULL,
  eps = 10^(-4),
  maxitr = 100,
  init = NULL
)
}
\arguments{
\item{X}{a matrix of size n by p containing n observations an p variables}

\item{l}{penalization parameter}

\item{m1}{the node at which the first partition occurs}

\item{m2}{the node at which the second partition occurs}

\item{m3}{the node at which the third partition occurs}

\item{m4}{the node at which the fourth partition occurs}

\item{m5}{the node at which the fifth partition occurs}

\item{m6}{the node at which the sixth partition occurs}

\item{m7}{the node at which the seventh partition occurs}

\item{m8}{the node at which the eighth partition occurs}

\item{m9}{the node at which the ninth partition occurs}

\item{eps}{tolerance parameter to decide whether algorithm has converved or not}

\item{maxitr}{maximum number of iterations to run before returning output}

\item{init}{initial estimate of graph adjacency B}
}
\value{
graph adjacency B
}
\description{
Estimates an adjacencey matrix for a DAG based on l1 penalized negative likelihood minimization given a partitioning of the nodes into two groups
}
\examples{
partial9(X = X, l = 2, m1 = 12, m2 = 24, m3 = 36, m4 = 48, m5 = 60, m6 = 72, m7 = 84, m8 = 96)

}
